+++ NEWS.md
@@ -1,5 +1,5 @@
 # censusapi 0.4.0
-* Adds support for NAICS code arguments used in [Business Patterns](https://www.census.gov/data/developers/data-sets/cbp-nonemp-zbp.html) APIs and [Economic Census](https://www.census.gov/data/developers/data-sets/economic-census.html) APIs.
+* Adds support for NAICS code arguments used in [Business Patterns](https://www.census.gov/data/developers/data-sets/cbp-nonemp-zbp.html) APIs, [Economic Census](https://www.census.gov/data/developers/data-sets/economic-census.html) APIs, and [Annual Survey of Manufactures](https://www.census.gov/data/developers/data-sets/Annual-Survey-of-Manufactures.html) APIs.
 
 # censusapi 0.3.0
 * Does not convert ACS annotation flag variables into numeric columns.